+++ Source/SwiftLintFrameworkTests/RulesTests.swift
@@ -91,6 +91,10 @@ class RulesTests: XCTestCase {
         verifyRule(OperatorFunctionWhitespaceRule.description)
     }
 
+    func testFunctionParameterCountRule() {
+        verifyRule(FunctionParameterCountRule.description)
+    }
+
     func testReturnArrowWhitespace() {
         verifyRule(ReturnArrowWhitespaceRule.description)
     }