+++ CHANGELOG.md
@@ -19,6 +19,9 @@
 * Improve performance of `ColonRule`.  
   [Norio Nomura](https://github.com/norio-nomura)
 
+* Add `FunctionParameterCountRule`.  
+  [Denis Lebedev](https://github.com/garnett)
+
 ##### Bug Fixes
 
 * None.